+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### 🎨 Enhancements
 
 - [#915](https://github.com/estruyf/vscode-front-matter/issues/915): Added a new setting `frontMatter.panel.openOnSupportedFile` which allows you to open the panel view on supported files
+- [#921](https://github.com/estruyf/vscode-front-matter/issues/921): Improve the filename sanitization
 
 ### ⚡️ Optimizations